fix: Consolidate config utils into a single common util
