--- internal/signer/darwin/util/test_data/certificate_config.json
@@ -1,8 +1,0 @@
-{
-  "cert_configs": {
-    "macos_keychain": {
-      "issuer": "Google Endpoint Verification"
-    }
-  }
-}
-
