--- internal/signer/linux/util/test_data/certificate_config.json
@@ -1,10 +1,0 @@
-{
-  "cert_configs": {
-    "pkcs11": {
-      "slot": "0x1739427",
-      "label": "gecc",
-      "user_pin": "0000",
-      "module": "pkcs11_module.so"
-    }
-  }
-}
